client: Allow passing custom download URLs to the download API
The download URL is the url returned in the downloader callback,
depicting the location of the downloader's job result for the client to
go and fetch it.

Up to now the downloader server added a default download URL to all of
the callbacks assuming that all of its jobs result in the same storage.

Now that downloader supports passing custom downloadURL per job we need
to support it on the client side also.

This commit also adds a new attribute in callbacks, named
`download_url_scheme` to be able to distinguish what kind is the backend
storage (e.g. http, https, s3)